Refactor React integration by replacing the old react-bridge with new functions for initialization and cleanup.
